+++ fast64_internal/oot/scene/exporter/to_c/room_commands.py
@@ -1,4 +1,4 @@
-from ....utility import CData
+from .....utility import CData
 
 
 def cmdEchoSettings(room, header, cmdCount):
