+++ fast64_internal/oot/scene/exporter/to_c/scene_header.py
@@ -1,6 +1,6 @@
-from ....utility import CData
-from ....f3d.f3d_gbi import ScrollMethod, TextureExportSettings
-from ...oot_f3d_writer import OOTGfxFormatter
+from .....utility import CData
+from .....f3d.f3d_gbi import ScrollMethod, TextureExportSettings
+from ....oot_f3d_writer import OOTGfxFormatter
 from .scene_pathways import ootPathListToC
 from .actor import ootTransitionActorListToC, ootStartPositionListToC, ootEntranceListToC
 from .scene_commands import ootSceneCommandsToC
