+++ fast64_internal/oot/scene/exporter/to_c/scene_pathways.py
@@ -1,4 +1,4 @@
-from ....utility import CData
+from .....utility import CData
 
 
 def ootPathToC(path, headerIndex: int, index: int):
